Assignment completed
Completed parallel-add-list and ran on Palmetto with 2, 4, 8, 16 and 32 processors.

To run on Palmetto you must have a copy of the parallel-add-list executable in the root directory (cd ~). Then run `make pbs` and it'll run for 2, 4, 8, 16 and 32 processors and then output .o###### files with the result.

Also able to run it locally using `mpirun -n [#] ./parallel-add-list`